[ld][dl] Add initial loadable TLS module for dl-load-tests
To fully test that dlclose doesn't cause problems with accessing TLS
variables through the DTV, we need to load an initial module before the
various tls-*dep modules. This patch adds that module to lib ld, and
uses it in both the tests for the fast and slow paths.

Fixed: 376130102
Change-Id: I75551c9b9fbea9841fa1077c5ec8628186cb8ed9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147006
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Paul Kirth <paulkirth@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>