[netstack-migration] Schedule Collaborative Reboot
When stack_migration observes a change to the automated/user setting
that is different from its `current_boot` value, it schedules a
collaborative reboot.

Once the collaborative reboot is initiated by the product, the device
will reboot and start up with the automated setting. This effectively
allows the device to self apply changes in the Netstack version.

There's additional logic in place to cancel the collaborative reboot,
should the setting change again (back to the original value).

Bug: 380873116
Change-Id: Ic87278b30de3c941c54313e0422e9518d9277d60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193375
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>