[inspect] Improve Block API
These changes result in the removal of multiple runtime validations
which are lifted to be statically checked by the type system and as a
side effect the removal of various Results.

Improvements of 10% up to 70% in inspect writing benchmarks.

See:

- VIM3: https://logs.chromium.org/logs/fuchsia/buildbucket/cr-buildbucket/8724426674705389009/+/u/compare_perf_test_results_without_and_with_CL/stdout

- X64: https://logs.chromium.org/logs/fuchsia/buildbucket/cr-buildbucket/8724426674705389025/+/u/compare_perf_test_results_without_and_with_CL/stdout

Bug: 392965471
Change-Id: Ie7f7777dd7b16530c581ff734ff07de5cf4400a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193057
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>