[display][coordinator] Sync after ApplyConfig() in integration tests.
Currently, the ApplyConfig() FIDL API and its higher-level ApplyLayers()
are prone to racy usage, leading to flaky tests. For performance
reasons, ApplyConfig() does not wait for the FIDL call to reach the
Display Coordinator and be executed. SendVsyncFromDisplayEngine() does
not go across the same FIDL pipe, so the resulting VSync signal may
reach the Coordinator before the config is applied.

Integration tests currently use a LayerCount() helper (which reaches
into the Coordinator implementation) as a way to wait until a
configuration is applied by the Coordinator. However, the helper is
only reliable when waiting for a configuration with a different number
of layers from the previous configuration.

This CL aims to remove this source of flakiness, and uses the following
approach.

1. ApplyLayers() is modified to issue a GetLastConfigStamp() and return
   its value. The call goes across the same FIDL pipe as the
   ApplyConfig(), so it is guaranteed to complete afterwards.
   A future CL will switch the ApplyLayers() implementation to use
   ApplyConfig3(), at which point ConfigStamp values will be generated
   in tests.

2. Direct usage of ApplyConfig() is followed by a GetLastConfigStamp().
   A future CL will remove the direct usage, and consolidate the
   integration tests around ApplyLayers().

3. DisplayLayerCount() is removed. Using GetLastConfigStamp() is more
   reliable, and the tests don't need to reach inside the Coordinator's
   presentation state.

Bug: 388885807
Test: fx test display-coordinator-unittests --count 100
Change-Id: I8c207460be41be4a42e376066a7298a2e469886d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193939
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>