[windowed-stats] Introduce filter mapping to reactors.
This change introduces the `filter_map_data_record` reactor combinator.
This combinator is the first that affects control flow on the basis of
inputs (rather than outputs like `then`, `and`, etc.). Filter mapping
can be used to react only to particular data that may not always be
available. For example, an input event may have variants, some of which
are to be sampled and some of which are to be ignored.

Note that discarding data events does not affect system events, which
are still observed.

Test: unit tests
Change-Id: Ia23742c691e0bf203db5a054992a5b7638563834
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189412
Reviewed-by: Kiet Tran <kiettran@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>