[fdf][compat] Perform all dlopen and dlsym calls in the framework.
This patch adds a modules section to the program stanza of the
component manifest for drivers. This section provides an optional
list of additional modules to load in addition to the driver as well
as a list of symbols the primary driver would like to obtain from
those modules. The symbols are then populated into the existing
symbols field in the start args.

Additonally, in order to support a feature in the compat driver, we
also provide support for overriding dynamic libraries provided during
the dynamic linking phase of the extra modules. This allows us to
redirect libdriver.so to compat.so.

Lastly, the compat driver was updated to take advantage of this
feature and avoid calling dlopen and dlsym itself.

Bug: 360403906
Change-Id: I2c9c36bec926cf0c40333d4d2a6f07f1355bb878
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186157
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>