[intel-display][dp] Read E-EDID when a display is available.
Previously, when a DisplayPort display was available, the intel-display
driver sent the DDC I2C interface to the display coordinator,
and the coordinator sent I2C commands to read the extended EDID
bytes.

With this change, DpDisplay reads the extended EDID bytes when a
display is available, and sends the bytes to the coordinator.

Test: core.x64 booted on NUC11TNHv5 with Google P2718EC monitor
      (with E-EDID support).
Test: core.x64 booted on NUC11TNHv5 with HP ZR30w monitor (no
      E-EDID support).
Test: core.x64 booted on NUC11TNHv5 with DP headless plug (no
      E-EDID support).
Bug: 390552176
Change-Id: I329f70a29cc6921596ecf9ba63119b989aea2fcb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192957
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>