[uart] null::Driver::config_name() to kConfigName
Move from constexpr function to constexpr variable.
This was a missed update from previous patches, and since it was never
referenced anywhere it still built and passed tests.

Now the body of Driver references this variable, instead of hard-coding
the string in multiple places.

Change-Id: I02540e7349bafc91892d3a0beebd3450ee1071af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193941
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>