[boot-shim] Introduce fallback for console arguments.
As a way to bridge the gap with linux bootloaders, we introduce a
fallback that interprets console=ttyXX arguments. For the time being we
take the index and assume it's the i-th non disabled uart in a
devicetree.

For x86 the index maps to the proper COM port, and the legacy
configurations are encoded in the table.

Change-Id: I484e6c47e413b304fdd1cec5958d662839933d52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120687
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>