[fxfs][fshost] Add support for per-container volumes
fshost now serves the fshost.StarnixVolumeProvider protocol, which
supports creating and/or mounting and unmounting a new volume. The
name of this starnix volume is defined by the product and plumbed
through to fshost via structured config. fshost no longer serves the
CryptManagement protocol.

The StarnixVolumeProvider currently only supports vending a single
volume, which is consumed by the main Starnix container. Expanding the
usage to support multiple volumes requires significant refactoring and
renaming of the StarnixVolume storage_config field.

The reason we are defining the Starnix volume name via a product config
as opposed to hardcoding a name in fshost is so that we can use a name
that we wouldn't have to change once we start supporting more volumes.

Bug: b/389783709
Change-Id: I5a31eb9b9a81b57daf7852ef9e8bb08a7b2f3f70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181143
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
API-Review: Aaron Wood <aaronwood@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>