[display][coordinator] Simplify integration tests opening sequence.
This CL removes the term "active client" from integration tests.  This
term was used to refer to two different concepts:

1. The Display Coordinator client that owns the displays has its
   configurations applied. Confusingly, the Coordinator's implementation
   uses `active_client_`, and the tests have been using
   IsActiveClient() for this concept.

2. VSync events are sent to the client that applied the displayed
   configuration. Comments have been using the term "active client"
   for this as well.

This CL replaces IsActiveClient(), which reached into the Display
Coordinator, with waiting for ownership notifications from the Display
Coordinator's FIDL API.

This CL modifies tests to apply a configuration (and wait for it to be
applied) when they expect VSync events. Waiting for a configuration to
be applied ensures the delivery of the SetVsyncEventDelivery() call
issued before applying the configuration.

Last, this CL removes the term "active" from comments.

Bug: 388885807
Test: fx test //src/graphics/display/drivers/coordinator
Change-Id: Id663599090e23d7f987b7ba8ad970ffb538e0bfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196697
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>