[kernel][lockup] Refactor mp_all_cpu_startup_sync_hook
Refactor mp_all_cpu_startup_sync_hook to use the new
CanDumpRegistersAndBacktrace function and expand on rationale for why
it's OK to panic there on production builds.

On production builds, if one or more secondaries fail to start up,
we'll attempt to dump diagnostics from the CPUs that failed to
check-in and then panic.  On development builds, we'll emit an oops
instead.  See the comments in the code for the rationale.

Change x86's DumpRegistersAndBacktrace implementation to not require
that interrupts are disabled.  This fixes a potential bug where we
might otherwise recursively panic.

Manually tested by hacking up the secondary startup to force them to
fail to check in, then verifying that we oops or panic.

Change-Id: Ie1f07e64daeae3f4c863a86ec0d5520932ba56be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194248
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>