[starnix] Remove starnix.runner from SDK partner_internal
The unstable FIDL library "fuchsia.starnix.runner" was in the
"partner_internal" category because `ffx_starnix` tool is in the
"partner" category. Use of this library by that tool was removed in
https://fxrev.dev/1170856.

This addresses one instance of an exception to "partner_internal"
stability rules.

Bug: 378470841, 364727245
Change-Id: I2b87ff7f45affb23fc4e4a24f9f19933b35f1242
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196796
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>