[userboot] VmoBuffer resize allocates page aligned sizes.
When reaching capacity this buffer would attempt to resize, up to the
minimum size to fit the string being written. Unfortunately, this was a
guaranteed fail, since in order to succeed the new size must be page
aligned.

Now we round up the size up to the nearest page boundary that contains
the appended string.

Fixed: 384793954
Change-Id: I41a9709089ba0a3e8fa7b0de4325d2b3dd120224
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192155
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>