[kernel][crashlog] Add runtime to RAM crashlog
This change adds the runtime field to the RAM mappable crashlog's
header. This required the creation of new ram_crashlog_v1_t and
ram_crashlog_header_v1_t structures, as the addition of this field
must be done in a backward compatible way with previous versions of the
crashlog.

Note that previous versions of the kernel that use v0 crashlogs will
not be able to recover v1 crashlogs, but this is not an issue because
we do not support downgrade OTAs.

Bug: 371420920
Change-Id: I9ad5f6621601166cecfc2038ae5afc1c6dd8dee7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179713
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>