[bazel] add self-test to refresh-bazel-compdb
This CL adds an option to run a self-test on the commands created when
running `fx refresh-bazel-compdb`. The self-test ensures that the
compile commands are valid.

This check is not intended to be run as part of CQ due to the amount of
time it adds. The purpose of this command is to help with debugging.

Bug: 392101170
Change-Id: I85406c4dab709714be6409fcb9eba23eae1ce032
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193397
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>