[time][inspect] Corrects the way inspect properties are recorded
The old code used the `record_*` API which churns through inspect
memory. This is incorrect and not needed.

This change moves the creation of inspect prpoperties out of the
loop, ensuring they are rewritten, and not recreated on each
loop iteration.

While I was there, I reorganized the properties a bit as well.
Since no tooling depends on this data schema just yet, it is
low risk to change.

Bug: 393633010
Change-Id: I1f93170a64e6e09eeb17011f5f5ca2d601325db8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196286
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>