[time] Wire up all the canonical targets
I noticed that 'persistence' was not getting tested, which led to
this.

In addition I had to change the name of the internal timekeeper
crate `persistence`, because its unit test name clashed with
the unit test name of an unrelated library.

Change-Id: If0e088849c83240d50ec71bacd1be2896e45bdd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1197615
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>