[kernel][vm] Mark modified on partial map
Ensures that a VMO is marked as modified even if only a subset of the
pages of a page fault are successfully mapped.

This has no present functional changes as multi page faults are not yet
exposed.

Change-Id: I14164be292ba7a152ccb1d84a2cc961e8ce28090
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196437
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>