[testrunner] Fix profile merging when no debuginfod servers are set.
If the env var where we provide the debuginfod server URL is empty,
strings.Split() returns a list of one element which is the empty string.
So the covargs merge function was treating that as a nonempty debuginfod
server and trying to use debuginfod to merge and then failing.

Bug: 394129919
Change-Id: Ib3d89a3b3482307addcf5add3472b02ea2e0b233
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198663
Reviewed-by: Danielle Kay <danikay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>