[docs] make tracing duration examples more concrete
I was confused by the `var`-iables in the example,
because the variable name didn't match the screenshot.

To eliminate that confusion for myself (and hopefully
others), this patch replaces the `var`-iables with
literal text that matches the example.

Note that while examining the code for the `var`-iables,
I realized that some of the C and C++ examples had an
unnecessary C-string delimited. So I removed those along
the way.

Change-Id: Ib80da47902ac31e2d3c21bd66123dd5be7d47c32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196287
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>