[ffx][fho] Removing reference to TargetInfoQuery
This removes the reference to TargetInfoQuery in the DeviceLookup trait.

All the usages of this trait used Option<String> to create a
TargetInfoQuery, so the trait was updated to use Option<String> and
the underlying implementation now accepts `impl Into<TargetInfoQuery>`
to localize the usage of this data type.

This is part of the refactoring of fho to remove all target and FIDL
dependencies so non-device related tools will build faster.

Change-Id: Ib451b8ee0d5e17c441f68eb41e41f781f6df2f13
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198425
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>