[ffx][fho]Removing TargetInfo reference from DeviceLookup
This removes the TargetInfo reference in the DeviceLookup trait
and replaces it with a Trait defined in fho.

This trait is implemented by TargetInfoHolder in target_holders.

This is done as part of the effort to remove fidl dependencies on
the fho crate in order to improve build times of non-device subtools.

Change-Id: Ice1d6949f22bbaf91ec82fc43af3274cab0e057e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198576
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>