[rbe] Swallow remote command output if only being used for cache hits.
With the remote_cache_only RBE mode we only use the remote
command result if it was for a cache hit, so there's no need to print
the output when the remote command fails but the local command succeeds.

This provides the expected DX when using deny_warnings=false along with
rbe_mode=remote_cache_only when there are warnings present in the current
checkout. The remote build will fail but that's expected and we shouldn't
print those errors.

Change-Id: If54278be0e0eb07825375ab6b6f0178f94a1edfa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199250
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>