[starnix][range_map] Remove iter_mut() method
This method breaks the invariants of the range map because if the client
modifies adjacent values in the range map, those ranges ranges should be
merged but won't be. Also, this method is expensive to implement when
backing the range map with a copy-on-write data structure because
mutable iteration implies a write to the entire data structure.

Change-Id: I205476052815c196988ec2a0937ee7ae0ca0e252
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198942
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>