[devfs][cf] Add ServiceMemberWatcher
Adding a way to watch a service directory for
instances of that service, and then return a
client end to a member protocol interface
that is part of that service.

This is specialized for the case where there is
only one interface in a service that a client would
be interested in connecting to.

ServiceMemberWatcher uses ServiceWatcher for the parsing
of the directory->Watch command.
ServiceWatcher is modified to increase ergonomics.
ServiceWatcher could now be used as a fallback for
services where multiple interfaces are provided.

bug: 361633233

Testing: added test to incoming_integration_test

Change-Id: Id052fa14e1d3124b6a1e99775d2116453d352352
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125747
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Garratt Gallagher <garratt@google.com>