[zxtest] For streams, conditionally allocate the assertion
The approach here is much more in line with what vanilla zxtest does for
evaluating a predicate.

(1) Delegate to a template function.
(2) On failure, generate information and return.

For streams, we follow a similar approach here, where we delegate to a
template function, which will evaluate the predicate. If the predicate
is true (e.g. success) we return a nullptr, otherwise we return a
std::unique_ptr<StreamableBase> that will contain the right information
and allow for streaming data into it. The lifetime of this unique_ptr is
tied to the if/else branch introduced by the macro.

Before we were unconditionally instantiating ostringstream, and 2
fbl::String (actual_symbol and expected_symbol) on every call.

This approach seems much simpler.

Fixed: 394101741
Change-Id: I9b944b014346058daacce9e21eb997934f206baf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198573
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>