[fidl][cpp] Stop reformatting golden files
We currently reformat golden files so that changes to the clang-format
implementation and rules don't result in golden failures. This is
taking 7-10 minutes of CPU time, though it's parallelizable so in
practice it only adds a minute and a half to local builds.

After this change lands any clang-format changes will need to be done
atomically with regenerating goldens (with `fx check-goldens`). This has
the advantage of allowing us to review the impact of updating
clang-format as part of landing it.

Bug: 393476961
Change-Id: I7f9d11732ea9317cfd621d847aa853661afc7716
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199485
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>