[kernel] Add a runtime flag to control kernel PCI logic on arm64
This removes the WITH_KERNEL_PCIE #ifdef and adds a boot option that
controls the kernel mode PCI initialization and system calls. This
option is defined in the kernel.arm64 namespace and is only checked on
aarch64. User space PCI handling has been enabled already on x86_64 and
RISC-V and works on some arm64 platforms.

On x86_64 and RISC-V this change compiles the PCI support code into the
kernel. This code is disabled at runtime.

On aarch64, the PCI code is compiled in (as it was before) and is now
guarded by a default-true runtime flag. We can disable this flag in
configurations that do not rely on the kernel mode support and remove it
once we no longer need it. Concretely, currently the qemu-arm64
configuration only works with kernel PCI support.

Bug: 42182308
Change-Id: I22006c0b8f60b067d4840888115a28d2ea92f3db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195575
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>