[vfs] Reduce the size of spawned futures
The size of the wrapping future created in ExecutionScope::spawn was
twice the size of the task passed into it. This was because the wrapping
future was capturing the task in an upvar and then moving it to a local
when polling it requiring space for 2 copies of the task.
https://github.com/rust-lang/rust/issues/108906

Writing our own future prevents duplicating the space requirements. The
task spawned in VolumesDirectory::add_directory_entry is now 15312
bytes, down from 30640 bytes.

Bug: b/393365596
Change-Id: Ib90a3ad99d42bd937effcac6bfa39d780ef188f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198686
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Chris Suter <csuter@google.com>