[tee] Partially implement TEE_CheckMemoryAccessRights
This provides a partial implementation of the
TEE_CheckMemoryAccessRights function. This can be used to check if an
access to a particular address range will fault as well as to query for
exclusive access. This implements the former check. To check for
exclusive access, we'll need to perform more tracking of mappings
created within the runtime - in particular we'll want to track memory
reference mappings.

Bug: 384584494
Change-Id: I10b2d7c0526e6d65324574708f9933b7f9eca14e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175104
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>