[fxfs] Add support to create unnamed temporary files
An unnamed temporary file is first created in the graveyard and marked
to be purged. It needs to be marked to be purged so that we can reclaim
back any storage used once the open counts drops to 0.

This change also changed `open_count` to change one of the top bits to
indicate if the file is an unnamed temporary file. We need to be able to
identify the type of file it is on `link_into` to determine if the
object needs to be removed from the graveyard (if it is an unnamed
temporary file). As we now have two bits in `open_count` that are used
to represent information about the file object, refactor `open_count`
such that it will be wrapped in a struct and we use helper functions to
get open counts and other information.

Bug: 293950309
Test: fx test fxfs-platform-test storage_conformance_test
Change-Id: I16030179623a2ec54bc7344c786c8419585b5eb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177512
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>