[perfcompare] Relax sample size check to not reject flaky metrics
Stop requiring that the sample size for each metric match
"--expected_sample_size" exactly.  Allow the actual sample size to be
smaller than that.  Only complain if the actual sample size is larger.

The check turned out to be too strict, because it causes perfcompare
builds and sample builds to fail if they contain flaky metrics,
i.e. metrics that a test produces sometimes but not always.  Flaky
metrics aren't rejected in regular CI/CQ builds, which made
perfcompare and sample builds stricter than regular builds, which is
undesirable.

This check was originally added to reject duplicate metrics, for which
rejecting actual_size > expected_size is enough.  We don't need to
reject actual_size < expected_size for catching flaky metrics in
general, because that is done by the metric name expectation files,
which allow flaky metrics to be allowlisted as being flaky.

Bug: 391497045
Run-All-Tests: True
Change-Id: I62ec5c989b44ff1df63191ffde6c99ea57840a35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198585
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>