[build] skipprofile attribute for excluded funcs
We currently use -fprofile-list= flag to exclude
instrumenting files that are under //prebuilt
https://clang.llvm.org/docs/UsersManual.html#cmdoption-fprofile-list

We were using the older format that is defined in
https://clang.llvm.org/docs/SanitizerSpecialCaseList.html

When older format is used, Clang adds noprofile attribute
to a function that blocks inlining if the caller and callee
have different values of this attribute as explained in
https://llvm.org/docs/LangRef.html#function-attributes.

This patch switches to a new format to skip instrumenting
a specific path by using "source:path=skip". This format
adds skipprofile attribute instead of noprofile attribute
to the functions that are under the provided path. This
attribute does not restrict inlining between functions in
the same way that the noprofile attribute does. This patch
also clarifies the comments about the usage of this new
format.

Bug: 387503298
Change-Id: I0d77931df8c2dfa2f5faadcc22857b13b46c6fd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1197420
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>
Fuchsia-Auto-Submit: Gulfem Savrun Yeniceri <gulfem@google.com>
Commit-Queue: Gulfem Savrun Yeniceri <gulfem@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>