[sdk][component] Temporarily disable lsan for fdf_rust_component_tests
LSan was detecting a leak in this test with fxr/1197316 applied. That
particular change doesn't directly affect this component. It makes
improvements to logging so it's likely this leak was always around, but
we only just surfaced the diagnostic. We should fix the leak, but for
now we can temporarily disable the check for this test.

Bug: 394632612
Change-Id: Ife3d51d24229e0ea10afdb1c0e5faa510040be69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200444
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>