[netstack3] Straighten ICMP echo NATing
Most of the support was already in place, but we didn't have integration
tests running, and the forced directionality on request/reply with
panics on the rewrite was subject to real runtime panics if replies are
observed before requests in conntrack.

This CL enables all the tests that we can to put more load on ICMP Echo
NATing, and changes the conntrack strategy to use both the src and dst
port ids as an ICMP id. This gets ICMP echo NATing to a more reliable
place, without having to teach conntrack about directionality just yet.

Reply-tuple rewriting is not mod-walled into conntrack so we can enforce
the mapping per protocol.

Fixed: 380506614

Change-Id: I85bf483f07a7c5de7fd90db8aafc45c082d37f49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199346
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>