[test_pilot] error restructuring
The principal goal of this CL is to replace some strings that were used
to identify error sources with the source errors themselves. This posed
a challenge, because the widely-used error enum would no longer be Eq
or PartialEq when some source errors were referenced.

To make this tractable, the single error enum UsageError was factored
in a logical way into BuildError and UsageError. A BuildError was
defined to wrap UsageErrors. This approach organizes errors in a
reasonable way while still allowing tests to use Eq/PartialEq when
asserting error results.

Test: tests updated

Bug: 327640651
Change-Id: If453b7362f9e09f955f05d369ebface034ebc89f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200364
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>