[kernel][vm] Provide SoftFaultInRange
Adds new VmAspace fault method for hinting that a soft fault is being
resolved in the context of a larger operation, and forwards this as a
multi page fault request to the VmMapping.

Aside from test code nothing uses this new method and this has no
functional change.

Change-Id: Ia3805503e68b5964fac947923776718c0aa63f7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190445
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>