[ffx] Fixes no-daemon regression for ffx-strict.
This fixes a regression in ffx-strict that causes the daemon to start
even when using ffx-strict.

In order to handle the dependency cycles, this moves `NetworkConnector`
into its own crate. It's not the best solution, but it's at least a
short path to fixing the overall regression.

Bug: 392190786
Change-Id: I31d0fb307b1a2d205a0d3c2e7a98953886e5646e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194249
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>