[ffxutil] refactor ffx invocation
There are a variety of functions in ffxutil/ffx.go for running ffx
(Run(), RunWithTimeout(), RunWithTargetAndTimeout(), etc.). The
ffx-strict work will add another set of parameters (required config
options) which is hard to implement with the current model, since the
options are not composable. This CL introduces an `ffxInvoker` struct
which holds the options, and methods that can run the command (run()) or
build the exec.Cmd (cmd()).

The existing functions are retained for compatibility and convenience,
for now, but some of the more complex and seldom used ones will
eventually be removed, with callers switching to the new interface.

Bug: 391964053
Change-Id: I5e96d2794ee6a6ee6bce45440d4276ab7d37d83e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192682
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>