[fxfs] Reduce the size futures with guards
Rust futures generated from async move {} blocks are at least double the
size of all captured variables:
https://github.com/rust-lang/rust/issues/108906

There are 3 locations in fxfs where a future and a guard object are
captured inside of another future with the purpose of running the
original future while the guard is held. Rolling our own future for
this specific case halves the size of the outer futures.

This reduces the size of the future spawned in
FxFile::create_connection_async for prefetch_keys from 10032 bytes to
5024 bytes.

Bug: b/393365596
Change-Id: Id46f9939191a5ec63076a4ad35d220a00831152f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198687
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>