[idk][build] Enable starting sub-builds earlier
Previously, the entire main build of the core IDK collection had to
complete before sub-builds could start. This included waiting for the
conformance tests and other targets that are not built in the
sub-builds.

To avoid this, create a peer IDK collection that only runs in the
sub-builds. The sub-build rule only depends on building this collection
in the main build, so the sub-builds can be started as soon as this
subset completes.

Bug: 393154095
Change-Id: Id6516d946a9afbce99af76013f47e1e625ba1810
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198647
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>