[network] Fix dns client shard typo
Fixes a typo introduced in I4318b6ace3e456929cb756c2a0f649f6bfc452ae in
the client shard of the DNS resolver. This was only used in a couple
tests, so the blast radius is minimal, but it means we were likely
missing logs for the DNS resolver for tests that were using it since
that CL landed.

Change-Id: Ic0e3218576de6aeeaf6b3135b279991cf4b63b0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201226
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>