[fake-display] Disallow releasing in-use resources.
The protocol between the Display Coordinator and engine drivers
specifies that the Coordinator will not release resources (images,
capture images) that are currently in-use (by the currently applied
configuration, or by an in-progress capture).

This CL adds a missing check enforcing this constraint in the
fake-display driver. Thanks to the check, incorrect releases will result
in an immediate crash, as opposed to having ServiceAnyCaptureRequest()
crash later. This change will hopefully make it easier to debug any
incorrect release.

This CL also replaces some comments with error logging. This also
makes it easier to debug errors, and aligns fake-display with other
drivers.

Last, this CL fixes an incorrect release in an unit test.

Bug: 394413629
Test: fx test //src/graphics/display
Change-Id: I1c6ebd2143f7b71ed4ea4ed8c5e65d5a6c8d453b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200746
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>