[ffx][writer] Refactor Format enum to command
This moves the Format enum to ffx/command:lib so the dependency
direction can be reversed and have writer depend on command.

This is not ideal, and perhaps there is a better crate for the Format
enum. It is however, good to switch the direction of the dependency
so the TryFromEnv implementations for the writers can be moved
to the writer crate and out of fho.

Bug: 392174596
Change-Id: I8eff7c0e08799f813655e49b4f26e6ea78335a20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201487
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>